LoadRGB4 color count is WORD. (AOS ignores upper WORD and some m68k programs leave garbage there)
git-svn-id: https://svn.aros.org/svn/aros/trunk@45275 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/graphics/loadrgb4.c
@@ -18,7 +18,7 @@
 /*  SYNOPSIS */
 	AROS_LHA(struct ViewPort *, vp, A0),
 	AROS_LHA(UWORD           *, colors, A1),
-	AROS_LHA(LONG             , count, D0),
+	AROS_LHA(WORD             , count, D0),
 
 /*  LOCATION */
 	struct GfxBase *, GfxBase, 32, Graphics)
@@ -54,7 +54,7 @@
 {
     AROS_LIBFUNC_INIT
 
-    LONG t;
+    WORD t;
 
     if (!vp)
         return;
